+++ CHANGELOG.md
@@ -11,11 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Send correct image size when uploading images on web
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- Remove timeline flickering due to switching between image fallback and thumbnail in reply elements
 
 ### Security
 
@@ -36,7 +40,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Show error to user when room list element action failed
-- Don't block opening room database
 - Forgetting room when declining invite to deleted room
 - Image detail view now displays image the first time it is opened
 - PDF details view displays valid PDFs again